Extract image fns to own helper utils file
