+++ cypress/utils/grid/image.js
@@ -1,0 +1,12 @@
+import { getDomain } from '../networking';
+
+// hash of the image in assets/GridmonTestImage.png
+export const imageHash = 'fe052e21c4bc4d76a2c841d97c5b2281cccd19bd';
+
+export function getImageHash() {
+  return imageHash;
+}
+
+export function getImageURL() {
+  return `${getDomain()}images/${getImageHash()}`;
+}
